[feature/multi_module] 각 레이어 모듈화 (#19)
