+++ app/src/main/AndroidManifest.xml
@@ -13,14 +13,14 @@
         android:theme="@style/Theme.EveryMeal_Android"
         tools:targetApi="31">
         <activity
-            android:name=".MainActivity"
+            android:name="com.everymeal.presentation.MainActivity"
             android:exported="true"
             android:label="@string/app_name"
             android:theme="@style/Theme.EveryMeal_Android">
 
         </activity>
         <activity
-            android:name=".SplashActivity"
+            android:name="com.everymeal.presentation.SplashActivity"
             android:exported="true"
             android:theme="@style/Theme.EveryMeal_Android">
             <intent-filter>
