--- app/src/main/java/com/everymeal/everymeal_android/SplashActivity.kt
@@ -1,27 +1,0 @@
-package com.everymeal.everymeal_android
-
-import android.os.Bundle
-import android.window.SplashScreen
-import androidx.activity.ComponentActivity
-import androidx.activity.compose.setContent
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.Surface
-import androidx.compose.ui.Modifier
-import com.everymeal.everymeal_android.ui.theme.EveryMeal_AndroidTheme
-
-class SplashActivity : ComponentActivity() {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContent {
-            EveryMeal_AndroidTheme {
-                Surface(
-                    modifier = Modifier.fillMaxSize(),
-                    color = MaterialTheme.colorScheme.background
-                ) {
-                    SplashScreen()
-                }
-            }
-        }
-    }
-}
