+++ data/build.gradle.kts
@@ -1,0 +1,29 @@
+@Suppress("DSL_SCOPE_VIOLATION") // TODO: Remove once KTIJ-19369 is fixed
+plugins {
+    id("com.android.library")
+    alias(libs.plugins.kotlinAndroid)
+}
+
+android {
+    namespace = "com.everymeal.data"
+    compileSdk = 33
+
+    defaultConfig {
+        minSdk = 28
+        targetSdk = 33
+
+        testInstrumentationRunner = "androidx.test.runner.AndroidJUnitRunner"
+    }
+
+    compileOptions {
+        sourceCompatibility = JavaVersion.VERSION_17
+        targetCompatibility = JavaVersion.VERSION_17
+    }
+    kotlinOptions {
+        jvmTarget = "17"
+    }
+}
+
+dependencies {
+
+}
