+++ presentation/build.gradle.kts
@@ -1,19 +1,16 @@
 @Suppress("DSL_SCOPE_VIOLATION") // TODO: Remove once KTIJ-19369 is fixed
 plugins {
-    alias(libs.plugins.androidApplication)
+    id("com.android.library")
     alias(libs.plugins.kotlinAndroid)
 }
 
 android {
-    namespace = "com.everymeal.everymeal_android"
+    namespace = "com.everymeal.presentation"
     compileSdk = 33
 
     defaultConfig {
-        applicationId = "com.everymeal.everymeal_android"
         minSdk = 28
         targetSdk = 33
-        versionCode = 1
-        versionName = "1.0"
 
         testInstrumentationRunner = "androidx.test.runner.AndroidJUnitRunner"
         vectorDrawables {
@@ -43,7 +40,7 @@ android {
     composeOptions {
         kotlinCompilerExtensionVersion = "1.4.7"
     }
-    packaging {
+    packagingOptions {
         resources {
             excludes += "/META-INF/{AL2.0,LGPL2.1}"
         }
@@ -51,9 +48,6 @@ android {
 }
 
 dependencies {
-    implementation(project(":presentation"))
-    implementation(project(":domain"))
-    implementation(project(":data"))
 
     implementation(libs.core.ktx)
     implementation(libs.lifecycle.runtime.ktx)
