+++ presentation/src/main/res/drawable/every_meal_logo_image.png
