+++ presentation/src/main/res/values/themes.xml
@@ -1,0 +1,5 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+
+    <style name="Theme.EveryMeal_Android" parent="android:Theme.Material.Light.NoActionBar" />
+</resources>
